feat(@angular-devkit/core): add JSON-schema feature
This is exported a "schema" from core.

This replaces the @ngtools/json-schema library from before. Benchmarks show
a 4x improvements on access, at the cost of having to eval a lot of code
when parsing the schema, in order to build the proxies and validators for
values.

This has support for future improvements such as subschema and schema
references, which were not possible with the old library. Additionally, the
hit for parsing can be reduced significantly later.

For now, compatibility with the CLI configuration was the main goal of this
feature. Parity with the old system has been achieved (except for .d.ts
serialization which is coming next), and new features have been added.