fix(@angular-devkit/build-optimizer): fix hoisted class declarations
The problem is that hoisted classes dont have the same name, so we need to change
the regex, and to check the end of it for the class name.

Also added a test to verify this works (based on RxJS issue).

Also refactored the complex regex to make a bit more sense.

Fixes #214.