fix(@angular-devkit/schematics): do not try to read the file on rename
Previously the logic for renaming was faulty in that it tried to create the new file
and delete the old one. If the file was on disk though it could not be read and threw
FileDoesNotExist. That logic is all wrong and the Virtual FS should simply try to
move the file using its host (which it now does properly).